requirements_web.txt
Flask==2.3.3
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.5.0
numpy<2.0
Pillow>=9.0.0